Create README.md
Both documents are part of the program code of the paper "Prediction of oil well production based on the time series model of optimized recursive neural network". 
The document "duolstm" is used to predict the future oil well production data by using LSTM neural network model with multi-factor analysis using five kinds of data that affect oil well production. The "india_gru" document is based on the five oil wells from India and the oil production history data of the field, using the GRU neural network model to predict the future oil production of the field. 
For more details, see the paper.